+++ modules/channel/src/test/scala/higherkindness/mu/rpc/channel/metrics/MonitoringChannelInterceptorTests.scala
@@ -20,8 +20,8 @@ package channel.metrics
 import cats.effect.{Clock, IO, Resource}
 import cats.syntax.apply._
 import higherkindness.mu.rpc.common._
+import higherkindness.mu.rpc.common.util.FakeClock
 import higherkindness.mu.rpc.internal.interceptors.GrpcMethodInfo
-import higherkindness.mu.rpc.internal.metrics.util.FakeClock
 import higherkindness.mu.rpc.internal.metrics.{MetricsOps, MetricsOpsRegister}
 import higherkindness.mu.rpc.protocol._
 import higherkindness.mu.rpc.testing.servers._
@@ -40,7 +40,8 @@ class MonitoringChannelInterceptorTests extends RpcBaseTestSuite {
     "generate the right metrics with proto" in {
       (for {
         metricsOps <- MetricsOpsRegister.build
-        _          <- makeProtoCalls(metricsOps)(_.serviceOp1(Request()))(protoRPCServiceImpl)
+        clock      <- FakeClock.build[IO]()
+        _          <- makeProtoCalls(metricsOps, clock)(_.serviceOp1(Request()))(protoRPCServiceImpl)
         assertion  <- checkCalls(metricsOps, List(serviceOp1Info))
       } yield assertion).unsafeRunSync()
     }
@@ -48,7 +49,8 @@ class MonitoringChannelInterceptorTests extends RpcBaseTestSuite {
     "generate the right metrics when calling multiple methods with proto" in {
       (for {
         metricsOps <- MetricsOpsRegister.build
-        _ <- makeProtoCalls(metricsOps) { client =>
+        clock      <- FakeClock.build[IO]()
+        _ <- makeProtoCalls(metricsOps, clock) { client =>
           client.serviceOp1(Request()) *> client.serviceOp2(Request())
         }(protoRPCServiceImpl)
         assertion <- checkCalls(metricsOps, List(serviceOp1Info, serviceOp2Info))
@@ -58,16 +60,18 @@ class MonitoringChannelInterceptorTests extends RpcBaseTestSuite {
     "generate the right metrics with proto when the server returns an error" in {
       (for {
         metricsOps <- MetricsOpsRegister.build
-        _          <- makeProtoCalls(metricsOps)(_.serviceOp1(Request()))(protoRPCServiceErrorImpl)
+        clock      <- FakeClock.build[IO]()
+        _          <- makeProtoCalls(metricsOps, clock)(_.serviceOp1(Request()))(protoRPCServiceErrorImpl)
         assertion  <- checkCalls(metricsOps, List(serviceOp1Info), serverError = true)
       } yield assertion).unsafeRunSync()
     }
 
   }
 
-  private[this] def makeProtoCalls[A](metricsOps: MetricsOps[IO])(f: ProtoRPCService[IO] => IO[A])(
+  private[this] def makeProtoCalls[A](metricsOps: MetricsOps[IO], clock: FakeClock[IO])(
+      f: ProtoRPCService[IO] => IO[A])(
       implicit H: ProtoRPCService[IO]): IO[Either[Throwable, A]] = {
-    implicit val clock: Clock[IO] = FakeClock[IO]
+    implicit val _: Clock[IO] = clock
     withServerChannel[IO](
       service = ProtoRPCService.bindService[IO],
       clientInterceptor = Some(MetricsChannelInterceptor(metricsOps, myClassifier)))