+++ modules/metrics/dropwizard/src/main/scala/higherkindness/mu/rpc/dropwizard/DropWizardMetrics.scala
@@ -14,20 +14,32 @@
  * limitations under the License.
  */
 
+package higherkindness.mu.rpc.dropwizard
+
+import java.util.concurrent.TimeUnit
+
+import cats.effect.Sync
+import com.codahale.metrics.MetricRegistry
+import higherkindness.mu.rpc.internal.interceptors.GrpcMethodInfo
+import higherkindness.mu.rpc.internal.metrics.MetricsOps
+import higherkindness.mu.rpc.internal.metrics.MetricsOps._
+import io.grpc.MethodDescriptor.MethodType._
+import io.grpc.Status
+
 /*
  * [[MetricsOps]] algebra able to record Dropwizard metrics.
  *
  * The list of registered metrics contains:
  *
  * {prefix}.{classifier}.active.calls - Counter
- * {prefix}.{classifier}.{serviceName}.{methodName}.{methodType}.messages.sent - Counter
- * {prefix}.{classifier}.{serviceName}.{methodName}.{methodType}.messages.received - Counter
+ * {prefix}.{classifier}.{serviceName}.{methodName}.messages.sent - Counter
+ * {prefix}.{classifier}.{serviceName}.{methodName}.messages.received - Counter
  * {prefix}.{classifier}.calls.header - Timer
  * {prefix}.{classifier}.calls.total - Timer
  * {prefix}.{classifier}.{methodType} - Timer
  * {prefix}.{classifier}.{status} - Timer
  *
- * {methodType} can be one of the following:
+ * {methodType} can be one of the following:
  *    - "unary-methods"
  *    - "client-streaming-methods"
  *    - "server-streaming-methods"
@@ -47,27 +59,14 @@
  *    - "unreachable-error-statuses"
  *
  */
-
-package metricsops
-
-import java.util.concurrent.TimeUnit
-
-import cats.effect.Sync
-import com.codahale.metrics.MetricRegistry
-import higherkindness.mu.rpc.internal.interceptors.GrpcMethodInfo
-import higherkindness.mu.rpc.internal.metrics.MetricsOps
-import higherkindness.mu.rpc.internal.metrics.MetricsOps._
-import io.grpc.MethodDescriptor.MethodType._
-import io.grpc.Status
-
-object DropWizardMetricsOps {
+object DropWizardMetrics {
 
   sealed trait GaugeType
   case object Timer   extends GaugeType
   case object Counter extends GaugeType
 
   def apply[F[_]](registry: MetricRegistry, prefix: String = "higherkinderness.mu")(
-      implicit F: Sync[F]) = new MetricsOps[F] {
+      implicit F: Sync[F]): MetricsOps[F] = new MetricsOps[F] {
 
     override def increaseActiveCalls(
         methodInfo: GrpcMethodInfo,