+++ modules/server/src/test/scala/higherkindness/mu/rpc/protocol/RPCJavaTimeTests.scala
@@ -25,13 +25,14 @@ import com.fortysevendeg.scalacheck.datetime.instances.jdk8._
 import com.fortysevendeg.scalacheck.datetime.GenDateTime._
 import com.fortysevendeg.scalacheck.datetime.jdk8.granularity.seconds
 import higherkindness.mu.rpc.common._
+import higherkindness.mu.rpc.internal.encoders.avro.javatime._
 import higherkindness.mu.rpc.internal.encoders.avro.javatime.marshallers._
 import higherkindness.mu.rpc.internal.encoders.pbd.javatime._
 import higherkindness.mu.rpc.protocol.Utils._
 import org.scalacheck.Arbitrary
 import org.scalatest._
 import org.scalacheck.Prop._
-import org.scalatest.prop.Checkers
+import org.scalatestplus.scalacheck.Checkers
 
 class RPCJavaTimeTests extends RpcBaseTestSuite with BeforeAndAfterAll with Checkers {